refactor: Location, Result 패키지 수정
