+++ src/test/java/domain/model/ResultTest.java
@@ -1,13 +1,12 @@
-package domain.vo;
+package domain.model;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ResultTest {
+public class ResultTest {
 
     @Test
     @DisplayName("글자 제한 실패 테스트")
@@ -35,19 +34,5 @@ class ResultTest {
         //then
         assertDoesNotThrow(() -> new Result(value1));
         assertDoesNotThrow(() -> new Result(value2));
-    }
-
-    @Test
-    @DisplayName("동등성 비교 테스트")
-    public void equalsTest() {
-        //given
-        String value = "value";
-        Result target = new Result(value);
-
-        //when
-        boolean result = target.equals(new Result("value"));
-
-        //then
-        assertThat(result).isTrue();
     }
 }
